feat: added reserveCharts list to cache
