+++ pkg/chartRepo/repository/ChartRefRepository.go
@@ -40,7 +40,6 @@ type ChartRefRepository interface {
 	FindByVersionAndName(name, version string) (*ChartRef, error)
 	CheckIfDataExists(location string) (bool, error)
 	FetchChart(name string) ([]*ChartRef, error)
-	MatchChartNamePrefix(name string) ([]*ChartRef, error)
 	FetchInfoOfChartConfiguredInApp(appId int) (*ChartRef, error)
 	FetchChartInfoByUploadFlag(userUploaded bool) ([]*ChartRef, error)
 }
@@ -115,18 +114,6 @@ func (impl ChartRefRepositoryImpl) FetchChart(name string) ([]*ChartRef, error) 
 	err := impl.dbConnection.
 		Model(&chartRefs).
 		Where("lower(name) = ?", strings.ToLower(name)).
-		Select()
-	if err != nil {
-		return nil, err
-	}
-	return chartRefs, err
-}
-
-func (impl ChartRefRepositoryImpl) MatchChartNamePrefix(name string) ([]*ChartRef, error) {
-	var chartRefs []*ChartRef
-	err := impl.dbConnection.
-		Model(&chartRefs).
-		Where("lower(name) LIKE ?", strings.ToLower(name)+"%").
 		Select()
 	if err != nil {
 		return nil, err
